refactor: move addressbar to separate component
